backend_python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.3
loguru==0.7.2
pytest==7.4.3
httpx==0.25.2